+++ examples/custom-components/src/custom-proto.ts
@@ -5,14 +5,15 @@ import {
   BaseClientComponentMessage,
   BaseComponentProto,
   AnyComponentProto,
+  BaseClientComponentCall,
 } from '@arcanejs/protocol';
 
 export const CUSTOM_NAMESPACE = 'custom';
 
 export type StopwatchComponentProto = BaseComponentProto<
-  typeof CUSTOM_NAMESPACE
+  typeof CUSTOM_NAMESPACE,
+  'stopwatch'
 > & {
-  component: 'stopwatch';
   state:
     | { type: 'stopped'; timeMillis: number }
     | { type: 'started'; startedAt: number };
@@ -32,6 +33,18 @@ export type StopwatchPressMessage = BaseClientComponentMessage<
   button: 'start-stop';
 };
 
+export type StopwatchGetTime = BaseClientComponentCall<
+  typeof CUSTOM_NAMESPACE,
+  'request-time'
+>;
+
+export interface CustomComponentCalls {
+  'request-time': {
+    call: StopwatchGetTime;
+    return: string;
+  };
+}
+
 export type CustomComponentMessage = StopwatchPressMessage;
 
 export const isCustomComponentMessage = <
@@ -42,3 +55,9 @@ export const isCustomComponentMessage = <
 ): message is CustomComponentMessage & { component: C } =>
   message.namespace === 'custom' &&
   (message as CustomComponentMessage).component === component;
+
+export const isCustomComponentCall = <A extends keyof CustomComponentCalls>(
+  call: BaseClientComponentCall<string, string>,
+  action: A,
+): call is CustomComponentCalls[A]['call'] =>
+  call.namespace === 'custom' && call.action === action;