+++ packages/protocol/src/index.ts
@@ -1,11 +1,15 @@
 import { Diff } from '@arcanejs/diff';
 
-export type BaseComponentProto<Namespace extends string> = {
+export type BaseComponentProto<
+  Namespace extends string,
+  Component extends string,
+> = {
   key: number;
   namespace: Namespace;
+  component: Component;
 };
 
-export type AnyComponentProto = BaseComponentProto<string>;
+export type AnyComponentProto = BaseComponentProto<string, string>;
 
 export type MetadataMessage = {
   type: 'metadata';
@@ -17,15 +21,34 @@ export type MetadataMessage = {
 
 export type SendTreeMsg = {
   type: 'tree-full';
-  root: BaseComponentProto<string>;
+  root: BaseComponentProto<string, string>;
 };
 
 export type UpdateTreeMsg = {
   type: 'tree-diff';
-  diff: Diff<BaseComponentProto<string>>;
+  diff: Diff<BaseComponentProto<string, string>>;
 };
 
-export type ServerMessage = MetadataMessage | SendTreeMsg | UpdateTreeMsg;
+export type CallResponseMsg<Namespace extends string, T> = {
+  type: 'call-response';
+  namespace: Namespace;
+  requestId: number;
+} & (
+  | {
+      success: true;
+      returnValue: T;
+    }
+  | {
+      success: false;
+      errorMessage: string;
+    }
+);
+
+export type ServerMessage =
+  | MetadataMessage
+  | SendTreeMsg
+  | UpdateTreeMsg
+  | CallResponseMsg<string, unknown>;
 
 export type BaseClientComponentMessage<Namespace extends string> = {
   type: 'component-message';
@@ -33,6 +56,45 @@ export type BaseClientComponentMessage<Namespace extends string> = {
   componentKey: number;
 };
 
+export type BaseClientComponentCall<
+  Namespace extends string,
+  Action extends string,
+> = {
+  type: 'component-call';
+  namespace: Namespace;
+  componentKey: number;
+  requestId: number;
+  action: Action;
+};
+
+export type BaseClientComponentCallPair<
+  Namespace extends string,
+  Action extends string,
+  Call extends BaseClientComponentCall<Namespace, Action>,
+  Return = unknown,
+> = {
+  call: Call;
+  return: Return;
+};
+
+// export type BaseClientCallResponsePairs<Namespace extends string, Actions extends string> = {
+//   [A in Actions]: BaseClientComponentCallPair<Namespace, A, BaseClientComponentCall<Namespace, A>, unknown>;
+// };
+
+export type CallForPair<
+  Namespace extends string,
+  Pairs,
+  Action extends string & keyof Pairs,
+> =
+  Pairs extends Record<Action, { call: infer R }>
+    ? Omit<R & BaseClientComponentCall<Namespace, Action>, 'requestId'>
+    : never;
+
+export type ReturnForPair<Pairs, Action extends string & keyof Pairs> =
+  Pairs extends Record<Action, { return: infer R }> ? R : never;
+
 export type AnyClientComponentMessage = BaseClientComponentMessage<string>;
 
-export type ClientMessage = AnyClientComponentMessage;
+export type AnyClientComponentCall = BaseClientComponentCall<string, string>;
+
+export type ClientMessage = AnyClientComponentMessage | AnyClientComponentCall;